refactor: Move delimiter check and simplify character skipping logic
Moved the delimiter checking conditional statements into its own function for clarity and reuse.
This is will keep it readable even when new types of delimiter checking logic is implemented.
Replaced the if/else structure of the character skipping and stopping logic with a while loop for improved readability, and the time complexity is also same.